+++ examples/vue/src/ServerContent.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts">
+console.log("Rendering ServerContent. This should only happen during build.");
+</script>
+
+<template>
+  <div class="box"><slot>Empty</slot></div>
+</template>