+++ packages/svelte/CHANGELOG.md
@@ -1,3 +1,24 @@
+# @capri-js/svelte 1.0.0 (2022-06-25)
+
+
+### Bug Fixes
+
+* add typings for virtual:capri-component ([97f76cf](https://github.com/capri-js/capri/commit/97f76cf1a12ffc7b77abe549558a7186924345f6))
+* make vite:dep-scan happy ([40fe6ba](https://github.com/capri-js/capri/commit/40fe6bae7bc86a3f6f5970202b0fc13c3a59e0bb))
+
+
+### Features
+
+* add Svelte support ([db56dee](https://github.com/capri-js/capri/commit/db56deea004349fd0d74194bd72db815fccba5d5))
+
+
+
+
+
+### Dependencies
+
+* **capri:** upgraded to 4.0.0
+
 # @capri-js/svelte 1.0.0-next.1 (2022-06-23)